Create frontend files and navbar component

+++ backend/controller/product.controller.js
@@ -48,11 +48,15 @@ export const updateProduct = async (req,res) =>{
 
 export const deleteProduct = async (req,res) => {
     const {id} = req.params;
+    if(!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({success: false, message: "Invalid Product Id"});
+    }
+    
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message: "Product Deleted"})
     } catch (error) {
         console.error("Error in Deleting Product", error.message);
-        res.status(404).json({success: false, message: "Product Not Found"})
+        res.status(500).json({success: false, message: "Server Error"})
     }
 };
